feat: implement modular configuration
Instead of manually choosing which modules we import, now every module can be
enabled using the options it defines. This makes it so that we can import
everything without any troubles and allows us to configure each system
independently from the others.

Additionally, home-manager modules have been moved to the modules folder so that
all modules are in the same directory.

Lastly, the helper function myLib.nixosSystem has been removed since it no
longer has any significant benefits and it creates unnecessary abstraction.