feat(darwin): add launchd agent to backup documents to iCloud
The Launchd agent is executed every day at 10:00.  It uses rsync to synchronise
the contents of the (local) Documents directory with the iCloud Documents
folder.

This, together with the previous commits setting up syncthing, allows us to keep
the Documents folder synced between linux and darwin systems, while also having
the folder backed up in iCloud at the same time.